+++ include/ws_functions/pwg.php
@@ -1,6 +1,7 @@
 <?php
 // +-----------------------------------------------------------------------+
-// | Piwigo - a PHP based photo gallery                                    |
+// | Phyxo - Another web based photo gallery                               |
+// | Copyright(C) 2014 Nicolas Roudaire           http://phyxo.nikrou.net/ |
 // +-----------------------------------------------------------------------+
 // | Copyright(C) 2008-2014 Piwigo Team                  http://piwigo.org |
 // | Copyright(C) 2003-2008 PhpWebGallery Team    http://phpwebgallery.net |
@@ -30,106 +31,84 @@
  *    @option int max_urls
  *    @option int prev_page (optional)
  */
-function ws_getMissingDerivatives($params, &$service)
-{
-  global $conf;
+function ws_getMissingDerivatives($params, &$service) {
+    global $conf;
 
-  if (empty($params['types']))
-  {
-    $types = array_keys(ImageStdParams::get_defined_type_map());
-  }
-  else
-  {
-    $types = array_intersect(array_keys(ImageStdParams::get_defined_type_map()), $params['types']);
-    if (count($types)==0)
-    {
-      return new PwgError(WS_ERR_INVALID_PARAM, "Invalid types");
-    }
-  }
-
-  $max_urls = $params['max_urls'];
-  $query = 'SELECT MAX(id)+1, COUNT(*) FROM '. IMAGES_TABLE .';';
-  list($max_id, $image_count) = pwg_db_fetch_row(pwg_query($query));
-
-  if (0 == $image_count)
-  {
-    return array();
-  }
-
-  $start_id = $params['prev_page'];
-  if ($start_id<=0)
-  {
-    $start_id = $max_id;
-  }
-
-  $uid = '&b='.time();
-
-  $conf['question_mark_in_urls'] = $conf['php_extension_in_urls'] = true;
-  $conf['derivative_url_style'] = 2; //script
-
-  $qlimit = min(5000, ceil(max($image_count/500, $max_urls/count($types))));
-  $where_clauses = ws_std_image_sql_filter( $params, '' );
-  $where_clauses[] = 'id<start_id';
-
-  if (!empty($params['ids']))
-  {
-    $where_clauses[] = 'id IN ('.implode(',',$params['ids']).')';
-  }
-
-  $query_model = '
-SELECT id, path, representative_ext, width, height, rotation
-  FROM '. IMAGES_TABLE .'
-  WHERE '. implode(' AND ', $where_clauses) .'
-  ORDER BY id DESC
-  LIMIT '. $qlimit .'
-;';
-
-  $urls = array();
-  do
-  {
-    $result = pwg_query(str_replace('start_id', $start_id, $query_model));
-    $is_last = pwg_db_num_rows($result) < $qlimit;
-
-    while ($row=pwg_db_fetch_assoc($result))
-    {
-      $start_id = $row['id'];
-      $src_image = new SrcImage($row);
-      if ($src_image->is_mimetype())
-      {
-        continue;
-      }
-
-      foreach($types as $type)
-      {
-        $derivative = new DerivativeImage($type, $src_image);
-        if ($type != $derivative->get_type())
-        {
-          continue;
+    if (empty($params['types'])) {
+        $types = array_keys(ImageStdParams::get_defined_type_map());
+    } else {
+        $types = array_intersect(array_keys(ImageStdParams::get_defined_type_map()), $params['types']);
+        if (count($types)==0) {
+            return new PwgError(WS_ERR_INVALID_PARAM, "Invalid types");
         }
-        if (@filemtime($derivative->get_path())===false)
-        {
-          $urls[] = $derivative->get_url().$uid;
+    }
+
+    $max_urls = $params['max_urls'];
+    $query = 'SELECT MAX(id)+1, COUNT(1) FROM '. IMAGES_TABLE .';';
+    list($max_id, $image_count) = pwg_db_fetch_row(pwg_query($query));
+
+    if (0 == $image_count) {
+        return array();
+    }
+
+    $start_id = $params['prev_page'];
+    if ($start_id<=0) {
+        $start_id = $max_id;
+    }
+
+    $uid = '&b='.time();
+
+    $conf['question_mark_in_urls'] = $conf['php_extension_in_urls'] = true;
+    $conf['derivative_url_style'] = 2; //script
+
+    $qlimit = min(5000, ceil(max($image_count/500, $max_urls/count($types))));
+    $where_clauses = ws_std_image_sql_filter( $params, '' );
+    $where_clauses[] = 'id<start_id';
+
+    if (!empty($params['ids'])) {
+        $where_clauses[] = 'id IN ('.implode(',',$params['ids']).')';
+    }
+
+    $query_model = 'SELECT id, path, representative_ext, width, height, rotation FROM '. IMAGES_TABLE;
+    $query_model .= ' WHERE '. implode(' AND ', $where_clauses) .' ORDER BY id DESC LIMIT '. $qlimit .';';
+
+    $urls = array();
+    do {
+        $result = pwg_query(str_replace('start_id', $start_id, $query_model));
+        $is_last = pwg_db_num_rows($result) < $qlimit;
+
+        while ($row=pwg_db_fetch_assoc($result)) {
+            $start_id = $row['id'];
+            $src_image = new SrcImage($row);
+            if ($src_image->is_mimetype()) {
+                continue;
+            }
+
+            foreach($types as $type) {
+                $derivative = new DerivativeImage($type, $src_image);
+                if ($type != $derivative->get_type()) {
+                    continue;
+                }
+                if (@filemtime($derivative->get_path())===false) {
+                    $urls[] = $derivative->get_url().$uid;
+                }
+            }
+
+            if (count($urls)>=$max_urls and !$is_last) {
+                break;
+            }
         }
-      }
+        if ($is_last) {
+            $start_id = 0;
+        }
+    } while (count($urls)<$max_urls and $start_id);
 
-      if (count($urls)>=$max_urls and !$is_last)
-      {
-        break;
-      }
+    $ret = array();
+    if ($start_id) {
+        $ret['next_page'] = $start_id;
     }
-    if ($is_last)
-    {
-      $start_id = 0;
-    }
-  } while (count($urls)<$max_urls and $start_id);
-
-  $ret = array();
-  if ($start_id)
-  {
-    $ret['next_page'] = $start_id;
-  }
-  $ret['urls'] = $urls;
-  return $ret;
+    $ret['urls'] = $urls;
+    return $ret;
 }
 
 /**
@@ -137,9 +116,8 @@ SELECT id, path, representative_ext, width, height, rotation
  * Returns Piwigo version
  * @param mixed[] $params
  */
-function ws_getVersion($params, &$service)
-{
-  return PHPWG_VERSION;
+function ws_getVersion($params, &$service) {
+    return PHPWG_VERSION;
 }
 
 /**
@@ -147,63 +125,59 @@ function ws_getVersion($params, &$service)
  * Returns general informations about the installation
  * @param mixed[] $params
  */
-function ws_getInfos($params, &$service)
-{
-  $infos['version'] = PHPWG_VERSION;
+function ws_getInfos($params, &$service) {
+    $infos['version'] = PHPWG_VERSION;
 
-  $query = 'SELECT COUNT(*) FROM '.IMAGES_TABLE.';';
-  list($infos['nb_elements']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.IMAGES_TABLE.';';
+    list($infos['nb_elements']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.CATEGORIES_TABLE.';';
-  list($infos['nb_categories']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.CATEGORIES_TABLE.';';
+    list($infos['nb_categories']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.CATEGORIES_TABLE.' WHERE dir IS NULL;';
-  list($infos['nb_virtual']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.CATEGORIES_TABLE.' WHERE dir IS NULL;';
+    list($infos['nb_virtual']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.CATEGORIES_TABLE.' WHERE dir IS NOT NULL;';
-  list($infos['nb_physical']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.CATEGORIES_TABLE.' WHERE dir IS NOT NULL;';
+    list($infos['nb_physical']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.IMAGE_CATEGORY_TABLE.';';
-  list($infos['nb_image_category']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.IMAGE_CATEGORY_TABLE.';';
+    list($infos['nb_image_category']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.TAGS_TABLE.';';
-  list($infos['nb_tags']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.TAGS_TABLE.';';
+    list($infos['nb_tags']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.IMAGE_TAG_TABLE.';';
-  list($infos['nb_image_tag']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.IMAGE_TAG_TABLE.';';
+    list($infos['nb_image_tag']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.USERS_TABLE.';';
-  list($infos['nb_users']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.USERS_TABLE.';';
+    list($infos['nb_users']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.GROUPS_TABLE.';';
-  list($infos['nb_groups']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.GROUPS_TABLE.';';
+    list($infos['nb_groups']) = pwg_db_fetch_row(pwg_query($query));
 
-  $query = 'SELECT COUNT(*) FROM '.COMMENTS_TABLE.';';
-  list($infos['nb_comments']) = pwg_db_fetch_row(pwg_query($query));
+    $query = 'SELECT COUNT(1) FROM '.COMMENTS_TABLE.';';
+    list($infos['nb_comments']) = pwg_db_fetch_row(pwg_query($query));
 
-  // first element
-  if ($infos['nb_elements'] > 0)
-  {
-    $query = 'SELECT MIN(date_available) FROM '.IMAGES_TABLE.';';
-    list($infos['first_date']) = pwg_db_fetch_row(pwg_query($query));
-  }
+    // first element
+    if ($infos['nb_elements'] > 0) {
+        $query = 'SELECT MIN(date_available) FROM '.IMAGES_TABLE.';';
+        list($infos['first_date']) = pwg_db_fetch_row(pwg_query($query));
+    }
 
-  // unvalidated comments
-  if ($infos['nb_comments'] > 0)
-  {
-    $query = 'SELECT COUNT(*) FROM '.COMMENTS_TABLE.' WHERE validated=\'false\';';
-    list($infos['nb_unvalidated_comments']) = pwg_db_fetch_row(pwg_query($query));
-  }
+    // unvalidated comments
+    if ($infos['nb_comments'] > 0) {
+        $query = 'SELECT COUNT(1) FROM '.COMMENTS_TABLE.' WHERE validated=\'false\';';
+        list($infos['nb_unvalidated_comments']) = pwg_db_fetch_row(pwg_query($query));
+    }
 
-  foreach ($infos as $name => $value)
-  {
-    $output[] = array(
-      'name' => $name,
-      'value' => $value,
-    );
-  }
+    foreach ($infos as $name => $value) {
+        $output[] = array(
+            'name' => $name,
+            'value' => $value,
+        );
+    }
 
-  return array('infos' => new PwgNamedArray($output, 'item'));
+    return array('infos' => new PwgNamedArray($output, 'item'));
 }
 
 /**
@@ -212,37 +186,30 @@ function ws_getInfos($params, &$service)
  * @param mixed[] $params
  *    @option int[] image_id
  */
-function ws_caddie_add($params, &$service)
-{
-  global $user;
+function ws_caddie_add($params, &$service) {
+    global $user;
 
-  $query = '
-SELECT id
-  FROM '. IMAGES_TABLE .'
-      LEFT JOIN '. CADDIE_TABLE .'
-      ON id=element_id AND user_id='. $user['id'] .'
-  WHERE id IN ('. implode(',',$params['image_id']) .')
-    AND element_id IS NULL
-;';
-  $result = array_from_query($query, 'id');
+    $query = 'SELECT id FROM '. IMAGES_TABLE;
+    $query .= ' LEFT JOIN '. CADDIE_TABLE .' ON id=element_id AND user_id='. $user['id'];
+    $query .= ' WHERE id IN ('. implode(',',$params['image_id']) .') AND element_id IS NULL;';
+    $result = array_from_query($query, 'id');
 
-  $datas = array();
-  foreach ($result as $id)
-  {
-    $datas[] = array(
-      'element_id' => $id,
-      'user_id' => $user['id'],
-      );
-  }
-  if (count($datas))
-  {
-    mass_inserts(
-      CADDIE_TABLE,
-      array('element_id','user_id'),
-      $datas
-      );
-  }
-  return count($datas);
+    $datas = array();
+    foreach ($result as $id) {
+        $datas[] = array(
+            'element_id' => $id,
+            'user_id' => $user['id'],
+        );
+    }
+    if (count($datas)) {
+        mass_inserts(
+            CADDIE_TABLE,
+            array('element_id','user_id'),
+            $datas
+        );
+    }
+
+    return count($datas);
 }
 
 /**
@@ -252,28 +219,23 @@ SELECT id
  *    @option int user_id
  *    @option string anonymous_id (optional)
  */
-function ws_rates_delete($params, &$service)
-{
-  $query = '
-DELETE FROM '. RATE_TABLE .'
-  WHERE user_id='. $params['user_id'];
+function ws_rates_delete($params, &$service) {
+    $query = 'DELETE FROM '. RATE_TABLE .' WHERE user_id='. $params['user_id'];
 
-  if (!empty($params['anonymous_id']))
-  {
-    $query .= ' AND anonymous_id=\''.$params['anonymous_id'].'\'';
-  }
-  if (!empty($params['image_id']))
-  {
-    $query .= ' AND element_id='.$params['image_id'];
-  }
+    if (!empty($params['anonymous_id'])) {
+        $query .= ' AND anonymous_id=\''.$params['anonymous_id'].'\'';
+    }
+    if (!empty($params['image_id'])) {
+        $query .= ' AND element_id='.$params['image_id'];
+    }
 
-  $changes = pwg_db_changes(pwg_query($query));
-  if ($changes)
-  {
-    include_once(PHPWG_ROOT_PATH.'include/functions_rate.inc.php');
-    update_rating_score();
-  }
-  return $changes;
+    $changes = pwg_db_changes(pwg_query($query));
+    if ($changes) {
+        include_once(PHPWG_ROOT_PATH.'include/functions_rate.inc.php');
+        update_rating_score();
+    }
+
+    return $changes;
 }
 
 /**
@@ -283,13 +245,12 @@ DELETE FROM '. RATE_TABLE .'
  *    @option string username
  *    @option string password
  */
-function ws_session_login($params, &$service)
-{
-  if (try_log_user($params['username'], $params['password'], false))
-  {
-    return true;
-  }
-  return new PwgError(999, 'Invalid username/password');
+function ws_session_login($params, &$service) {
+    if (try_log_user($params['username'], $params['password'], false)) {
+        return true;
+    }
+
+    return new PwgError(999, 'Invalid username/password');
 }
 
 
@@ -298,13 +259,12 @@ function ws_session_login($params, &$service)
  * Performs a logout
  * @param mixed[] $params
  */
-function ws_session_logout($params, &$service)
-{
-  if (!is_a_guest())
-  {
-    logout_user();
-  }
-  return true;
+function ws_session_logout($params, &$service) {
+    if (!is_a_guest()) {
+        logout_user();
+    }
+
+    return true;
 }
 
 /**
@@ -312,22 +272,18 @@ function ws_session_logout($params, &$service)
  * Returns info about the current user
  * @param mixed[] $params
  */
-function ws_session_getStatus($params, &$service)
-{
-  global $user;
+function ws_session_getStatus($params, &$service) {
+    global $user;
 
-  $res['username'] = is_a_guest() ? 'guest' : stripslashes($user['username']);
-  foreach ( array('status', 'theme', 'language') as $k )
-  {
-    $res[$k] = $user[$k];
-  }
-  $res['pwg_token'] = get_pwg_token();
-  $res['charset'] = get_pwg_charset();
+    $res['username'] = is_a_guest() ? 'guest' : stripslashes($user['username']);
+    foreach ( array('status', 'theme', 'language') as $k ) {
+        $res[$k] = $user[$k];
+    }
+    $res['pwg_token'] = get_pwg_token();
+    $res['charset'] = get_pwg_charset();
 
-  list($dbnow) = pwg_db_fetch_row(pwg_query('SELECT NOW();'));
-  $res['current_datetime'] = $dbnow;
-  $res['version'] = PHPWG_VERSION;
-  return $res;
+    list($dbnow) = pwg_db_fetch_row(pwg_query('SELECT NOW();'));
+    $res['current_datetime'] = $dbnow;
+    $res['version'] = PHPWG_VERSION;
+    return $res;
 }
-
-?>
